+++ tests/integration/relations/test_db_admin.py
@@ -1,152 +1,93 @@
+#!/usr/bin/env python3
 # Copyright 2022 Canonical Ltd.
 # See LICENSE file for licensing details.
-
-import asyncio
+import ast
+import json
 import logging
-from pathlib import Path
 
 import pytest
-import yaml
+from landscape_api.base import run_query
 from pytest_operator.plugin import OpsTest
 
-from tests.integration.helpers import helpers
+from tests.integration.helpers.helpers import (
+    deploy_and_relate_bundle_with_pgbouncer_bundle,
+    deploy_postgres_bundle,
+)
+from tests.integration.helpers.postgresql_helpers import (
+    check_database_users_existence,
+    check_databases_creation,
+)
 
 logger = logging.getLogger(__name__)
 
-METADATA = yaml.safe_load(Path("./metadata.yaml").read_text())
-PGB = METADATA["name"]
-PG = "postgresql"
-PSQL = "psql"
-APPS = [PG, PGB, PSQL]
-BUILT_CHARM = None
-
-# TODO update to use different charms for testing
-
-
-@pytest.mark.abort_on_fail
-@pytest.mark.legacy_relation
-async def test_create_db_admin_legacy_relation(ops_test: OpsTest):
-    """Test that the pgbouncer and postgres charms can relate to one another."""
-    # Build, deploy, and relate charms.
-    charm = await ops_test.build_charm(".")
-    async with ops_test.fast_forward():
-        await asyncio.gather(
-            ops_test.model.deploy(
-                charm,
-                application_name=PGB,
-            ),
-            ops_test.model.deploy(PG, channel="edge", trust=True, num_units=3),
-            # Deploy a psql client shell charm
-            ops_test.model.deploy("postgresql-charmers-postgresql-client", application_name=PSQL),
-        )
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(apps=[PG], status="active", timeout=1000),
-            ops_test.model.wait_for_idle(apps=[PGB, PSQL], status="blocked", timeout=1000),
-        )
-        await ops_test.model.add_relation(f"{PGB}:backend-database", f"{PG}:database"),
-
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(apps=[PG, PGB], status="active", timeout=1000),
-            ops_test.model.wait_for_idle(apps=[PSQL], status="blocked", timeout=1000),
-        )
-        await ops_test.model.add_relation(f"{PGB}:db-admin", f"{PSQL}:db"),
-        await ops_test.model.wait_for_idle(apps=APPS, status="active", timeout=1000)
-
-    unit = ops_test.model.units[f"{PGB}/0"]
-    cfg = await helpers.get_cfg(ops_test, unit.name)
-    assert "cli" in cfg["databases"].keys()
+HAPROXY_APP_NAME = "haproxy"
+LANDSCAPE_APP_NAME = "landscape-server"
+LANDSCAPE_SCALABLE_BUNDLE_NAME = "ch:landscape-scalable"
+RABBITMQ_APP_NAME = "rabbitmq-server"
+DATABASE_UNITS = 3
 
 
 @pytest.mark.legacy_relation
-async def test_add_replicas(ops_test: OpsTest):
-    await ops_test.model.applications[PSQL].add_units(count=2)
+async def test_landscape_scalable_bundle_db(ops_test: OpsTest, charm: str) -> None:
+    """Deploy Landscape Scalable Bundle to test the 'db-admin' relation."""
+    config = {
+        "extra-packages": "python-apt postgresql-contrib postgresql-.*-debversion postgresql-plpython.*"
+    }
+    await deploy_postgres_bundle(ops_test, pg_config=config)
+
     async with ops_test.fast_forward():
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[PG], status="active", timeout=1000, wait_for_exact_units=3
-            ),
-            ops_test.model.wait_for_idle(
-                apps=[PSQL], status="active", timeout=1000, wait_for_exact_units=3
-            ),
-            ops_test.model.wait_for_idle(apps=[PGB], status="active"),
+        ops_test.model.deploy()
+        # Deploy and test the Landscape Scalable bundle (using this  charm).
+        relation_id = await deploy_and_relate_bundle_with_pgbouncer_bundle(
+            ops_test, LANDSCAPE_SCALABLE_BUNDLE_NAME, LANDSCAPE_APP_NAME, relation_name="db-admin"
         )
-    unit = ops_test.model.units[f"{PGB}/0"]
-    cfg = await helpers.get_cfg(ops_test, unit.name)
-    expected_databases = ["cli", "cli_standby"]
-    for database in expected_databases:
-        assert database in cfg["databases"].keys()
+    await check_databases_creation(
+        ops_test,
+        [
+            "landscape-account-1",
+            "landscape-knowledge",
+            "landscape-main",
+            "landscape-package",
+            "landscape-resource-1",
+            "landscape-session",
+        ],
+    )
 
+    landscape_users = [f"relation-{relation_id}"]
 
-@pytest.mark.legacy_relation
-async def test_remove_db_admin_unit(ops_test: OpsTest):
-    await ops_test.model.destroy_unit("psql/1")
-    async with ops_test.fast_forward():
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[PSQL], status="active", timeout=1000, wait_for_exact_units=2
-            ),
-            ops_test.model.wait_for_idle(
-                apps=[PG, PGB],
-                status="active",
-                timeout=1000,
-            ),
-        )
+    await check_database_users_existence(ops_test, landscape_users, [])
 
+    # Configure and admin user in Landscape and get its API credentials.
+    unit = ops_test.model.applications[LANDSCAPE_APP_NAME].units[0]
+    action = await unit.run_action(
+        "bootstrap",
+        **{
+            "admin-email": "admin@canonical.com",
+            "admin-name": "Admin",
+            "admin-password": "test1234",
+        },
+    )
+    result = await action.wait()
+    credentials = ast.literal_eval(result.results["api-credentials"])
+    key = credentials["key"]
+    secret = credentials["secret"]
 
-@pytest.mark.legacy_relation
-async def test_remove_backend_unit(ops_test: OpsTest):
-    await ops_test.model.destroy_unit("postgresql/1")
-    async with ops_test.fast_forward():
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[PG], status="active", timeout=1000, wait_for_exact_units=2
-            ),
-            ops_test.model.wait_for_idle(apps=[PGB, PSQL], status="active", timeout=1000),
-        )
+    # Connect to the Landscape API through HAProxy and do some CRUD calls (without the update).
+    haproxy_unit = ops_test.model.applications[HAPROXY_APP_NAME].units[0]
+    api_uri = f"https://{haproxy_unit.public_address}/api/"
+    role_name = "User"
 
+    # Create a role and list the available roles later to check that the new one is there.
+    run_query(key, secret, "CreateRole", {"name": role_name}, api_uri, False)
+    api_response = run_query(key, secret, "GetRoles", {}, api_uri, False)
+    assert role_name in [user["name"] for user in json.loads(api_response)]
 
-@pytest.mark.legacy_relation
-async def test_remove_db_admin_leader(ops_test: OpsTest):
-    await ops_test.model.destroy_unit("psql/0")
-    async with ops_test.fast_forward():
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[PSQL], status="active", timeout=1000, wait_for_exact_units=1
-            ),
-            ops_test.model.wait_for_idle(
-                apps=[PG, PGB],
-                status="active",
-                timeout=1000,
-            ),
-        )
-    unit = ops_test.model.units[f"{PGB}/0"]
-    cfg = await helpers.get_cfg(ops_test, unit.name)
-    assert "cli" in cfg["databases"].keys()
+    # Remove the role and assert it isn't part of the roles list anymore.
+    run_query(key, secret, "RemoveRole", {"name": role_name}, api_uri, False)
+    api_response = run_query(key, secret, "GetRoles", {}, api_uri, False)
+    assert role_name not in [user["name"] for user in json.loads(api_response)]
 
-
-@pytest.mark.legacy_relation
-async def test_remove_backend_leader(ops_test: OpsTest):
-    await ops_test.model.destroy_unit("postgresql/0")
-    async with ops_test.fast_forward():
-        await asyncio.gather(
-            ops_test.model.wait_for_idle(
-                apps=[PG], status="active", timeout=1000, wait_for_exact_units=1
-            ),
-            ops_test.model.wait_for_idle(apps=[PGB, PSQL], status="active", timeout=1000),
-        )
-    unit = ops_test.model.units[f"{PGB}/0"]
-    cfg = await helpers.get_cfg(ops_test, unit.name)
-    assert "cli" in cfg["databases"].keys()
-
-
-@pytest.mark.legacy_relation
-async def test_remove_db_admin_legacy_relation(ops_test: OpsTest):
-    """Test that removing relations still works ok."""
-    await ops_test.model.applications[PGB].remove_relation(f"{PGB}:db-admin", f"{PSQL}:db")
-    await ops_test.model.applications[PSQL].remove()
-    async with ops_test.fast_forward():
-        await ops_test.model.wait_for_idle(apps=[PGB, PG], status="active", timeout=1000)
-
-    unit = ops_test.model.units[f"{PGB}/0"]
-    cfg = await helpers.get_cfg(ops_test, unit.name)
-    assert "cli" not in cfg["databases"].keys()
+    # Remove the applications from the bundle.
+    await ops_test.model.remove_application(LANDSCAPE_APP_NAME, block_until_done=True)
+    await ops_test.model.remove_application(HAPROXY_APP_NAME, block_until_done=True)
+    await ops_test.model.remove_application(RABBITMQ_APP_NAME, block_until_done=True)
